requirements.txt
discord.py>=2.3.0
feedparser>=6.0.0
aiohttp>=3.9.4
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
